restore tar extract functionality
This also does the download to a temp partial file.

If the user does not pass `-x`, we finish by renaming the temp file to
the target file.  This means that if pget is interrupted, we don't
leave an incomplete file with the target filename.

If the users does pass `-x`, we pass the temp file as input to
TarReader, then delete it at the end.