Implement download modes framework (#25)
Imeplement a download modes framework that allows specifying a mechanism
to download. This commit implements "buffer" and "tar-extract" which mirror
the original behaviors of pget. The '-x'/'--extract' option automatically
sets the `tar-extract` mode.

tar-extract mode simply wraps the buffer mode for now.

Behavior and CLI options are not modified in this commit