Implement Multifile
Implement the first pass at multifile command for pget.

This will take input on STDIN or a filename. If on STDIN EOF or an
empty line will trigger the start of downloading.

Logging will need updates as it is awful/almost unusable with multifile
as is. This will be done in a subsequent commit.

Total connection count (independent from per-host) is not currently
limited. This will be implenment in a subsequent commit.