Fix test flake
TestDownloadFilesFromHost was flaking.  This was because we were writing to
dummyMode.args and dummyMode.timesCalled in an unsynchronised way from two
different goroutines.  This meant that we could get lost updates.

I fixed this by adding a channel.  dummyMode.DownloadFile now sends its args to
a channel; once you're done with your test input, you call dummyMode.Args() to
read everything off the channel and put it into the slice as before.