Ensure a shared client is used
Ensure a shared client is used even in multifile mode. This prevents
a new client per file from being created. This is handled by instead
of passing in a config object to the mode, a client is passed into
the mode. The NewHTTPClient function now takes a single argument
of HTTPClientOpts which defines the options expected. These
then can/will be set by the calling CMD.

This pulls viper calls one layer higher up.