Fix --force
It should be safe to use `O_TRUNC` in all cases in the FileWriter
as we check earlier in the stack. If we get this far without checking
for existence and the use of the --force flag, it is likely a programming
error further up the stack from the consumer.