Change config vars
- make MaxConnPerHost globally persistent

We want to set this (and MaxIdleConnsPerHost) always.

It's now a hidden option, this isn't our primary mechanism for concurrency
control. It's there to: 1) protect nginx from too many connections, 2) support
keeping idle connections open.

- prefer --concurrency over --max-chunks

For multifile, we want to be able to download chunks from different files in
parallel.

- remove unused name MaxConcurrentFiles

Never referenced anywhere