Fix bug causing incorrect range-requests related to max chunks
When MaxChunks > fileSize / minChunkSize the number of chunks
was reset and due to an error in calculating the offset of the
start of remaining bytes being minChunkSize instead of ChunkSize
we would end up with files that had incorrect gaps in the
file structure where bytes equivalent to chunkSize - minChunkSize
would simply be skipped after the first minChunkSize block.

This commit corrects the issue by simplifying the means we use
to calculate chunkSize and number of chunks.

Closes: #73