Introduce null consumer and Hidden option to set consumer (#60)
* Implement GetConsumer in pkg/config

pkg/Config now has the helper function GetConsumer to map consumer
names to a valid consumer. This function explicitly calls viper (as
part of the config helpers). CMDs are expected to call config.GetConsumer()
and utilize the resulting consumer.Consumer for the Getter's "Consumer"
attribute.

* Implement NullWriter

Implement NullWriter and wire up config to utilize it for --output=null