Cleanup (#97)
* Eliminate global Host resolution map

Implement a struct that is attached to the transport for the dialcontext.
This struct contains the resolution override map. It is now expected
that commands will pass down the map as part of the Client Options.

* Move the .doRequest for a given chunk into the goroutine

Quick benchmark shows a 200-300MB/s increase on slower downloads
(e.g. from one provider to another) when not blocking the primary
execution block to make the request prior to downloading the chunk.