Implement consistent-hash fallback for files and chunks
This commit implements consistent hash fallback behavior when a
file's first request fails or a given chunk fails. This is handled
in the retryablehttp client's `RetryPolicy` function. In any case
that a fall-back is desired, the `RetryPolicy` will return a `false`
(meaning do not retry) and a special `ErrStrategyFallBack` that the
`Strategy` can act upon.

This commit exposes DoRequest as part of the Strategy Interface to
handle the fallback for a single chunk.

`RetryPolicy`'s additional logic above (and beyond) the
`DefaultRetryPolicy` will only trigger when using consistent-hash
strategy.