introduce bufferedReader
The idea of bufferedReader is to be a staging area where we can download request
bodies to.  It blocks Read() calls until the whole request has been downloaded.

The key thing this enables us to do is to get rid of the errGroup.Wait() call,
because the readers that we return are waiting for each chunk to be ready and we
no longer need to wait until the whole file has been downloaded to return from
Fetch().