we don't need to implement WriteTo
WriteTo can help avoid allocations where you call io.Copy(w, r); if the reader
implements WriteTo or the writer implements ReadFrom then data flows directly
from one to the other; if neither do then io.Copy needs to allocate temporary
staging buffers to copy the data across.

Right now we're not benefitting from this because we'd need to implement it in
chanMultiReader as well, and it's not trivial to implement there.

The thing is, this only matters for Consumers that do a copy to a writer that
doesn't implement ReadFrom.  We have:

- FileWriter, which writes to a *File, which implements ReadFrom
- TarExtractor, which feeds the Reader directly to tar.Reader
- NullWriter, which would in fact benefit (because io.Discard doesn't implement
  ReadFrom)

This doesn't seem worthwhile. Therefore, deleting this code.