Consistent hashing: consistently retry another host (#119)
* Consistent hashing: consistently retry another host

This implements a retry policy to consistently retry another host within the SRV
set if we get a failure talking to the original host.

We still use the original fallback strategy if this retry fails. (Should we?)

* extract consistent hashing code

This extracts the consistent hashing code into a new package and writes more
tests for it.

This removes a bunch of duplication from ConsistentHashingMode.

As a consequence this jumbles the consistent hashing algorithm (because I've
embedded the Key into a top-level CacheKey struct which includes Attempt for
retry support). At this stage this is safe because nothing live is using it.

* refactor

rename method

avoid implicit return

remove unneeded conditional on m.DomainsToCache

* lint