Improve PIDFile locking
Instead of using flock, use exclusive create. This method has
the blind spot that if the pid file has been left after a failed
run, or crashed process, it is not possible to automatically clean
up.

Likely it will make sense to add in an addtinal check "does pid exist
from file if file exists" and "verify if pid is a pget process". If
it's a stale pid we can safely remove and try the O_EXCL create...