Implement consistent-hash domains configurable
Implement making consistent-hash domains to cache configurable.
** THIS IS A BREAKING CHANGE **

Consistent-Hash now utilizes a path-based proxy system, prepending
the URI to the path prior to sending to the consistent-hash index
target. This URL manipulation does not change the hash value. Buckets
are maintained.

The prefixes allowed to be cached are provides exclusively as an ENV
var and is in the format of "scheme://host/prefix, scheme://host/prefix, ..."

Eventually these "must only be set via ENV" should move to a config
file.

The consistent hash services already understand the new path-based
proxy.