Improve consistent-hash testing with httpmock
Following a comment by @nickstenning [1], it's his fault.

This changes most (not all) consistent-hash tests to use httpmock to mock out
http requests.  This is helpful because standard httptest uses random ports to
listen on, which changes URL hash values and messes up the resulting consistent
hash behaviour. It's also helpful because I can rewrite some tests that relied
on the existence or otherwise of URLs on the public internet.

client.Options has been updated to allow a caller to pass in a Transport; I've
also made it clear which options are TransportOptions (and therefore will be
ignored if the caller passes their own Transport).

I write a couple of helper functions:

- rangeResponder is an httpmock.Responder that understands http range requests
  and responds accordingly
- fakeCacheHosts creates an httpmock.MockTransport with a set of pre-registered
  cache hosts that response to /hello.txt with a static file with contents that
  are different for each cache host.

[1]: https://github.com/replicate/pget/pull/155#discussion_r1488358253