Add multiReader
multiReader is a reader that implements the ReadAt functionality needed
for some future consumers (e.g. unzip). The multiReader at a basic level
consumes a mutltiChanReader via the NewmultiReader() function and
returns an io.ReaderAt implementation.

bufferedReader now has a .len() calculation that will report the
content-length once that header is received. Since we do not know the
actual content length until the download starts, there is a new signal
channel to indicate the download has started and allows us to read the
size of the bufferedReader.

This means that there is the real likelihood that reading from
multiReader may block more often than chanmultiReader. MultiReader may
be able to implement Seek() and other related functions for reading the
data out of strict order.