Reuse buffers where possible
This commit imeplemnts use of `sync.Pool` for the channle multireader
allowing pget to re-use already allocated buffers. This is intended to
be the first step to reduce overal ram utilization, as consumers can
start consuming before the entire file is downloaded. This means that
smaller numbers of buffers will be allocated.

NOTE: The buffer pool is on a per-file-download basis, this means that
multifile still has the potential to use significantly more memory than
single file mode.