WIP: fixed chunk size
We have a fair amount of complexity (EqualSplit etc) involved in dynamically
splitting files down into chunks based on available concurrency.

We could be less clever and use a fixed chunk size, with a fixed buffer size to
go with it.  This would dovetail nicely with a sync.Pool of fixed-size buffers,
and also with a configuration option to limit the total allocated
RAM (implemented as a semaphore guarding how many buffers can be fetched from
the sync.Pool).

This commit only does consistent hashing mode; we should do buffer mode as well
if we do this work.