BufferedReader Signal Channel -> sync.Cond
To cleanup the use of the signal channel and enable easy re-use of the
BufferedReader this commit converts the channel to a sync.Cond. This
allows a reset to simply call:

b.ready = false
b.buf.Reset()

It also eliminates the need to re-create the channel on a reset case.