Create Buffered Reader Pool
Supporting reusable readers will help us reduce memory needs of pget.

This commit implements the buffered reader pool and wires up CH and
buffer mode to use it to acquire the buffered readers. This lightens the
logic even further guaranteeing that we always use the same capacity for
all readers as it's defined in the  `New` function used by the
sync.Pool

This commit does not return buffers to the pool.