Return Buffered Readers to the Pool
When we're done with a bvuffered reader, call `.Close()` which will do
the reset and re-add to the pool. This change also pivots the
BufferedReader to a ReadCloser type interface.

This change follows through on the promise of allowing reuse of buffers,
minimizing allocations, and reducing overall memory requirements of
pget. This will help further limit OOM situations and lead towards a
future of also limiting maximum memory utilization of pget.