Fix tar-extract
Tar-extract had a bug with the zipSlip guard. This PR fixes so we can
extract files properly the issue was taking abs of the target multiple
times and not the destDir, stringified the paths were equal, but bytes
representation were different.

This commit does not include a regression test (yet).