Cleanup Aliasing
Aliases in Cobra and Viper are relatively complex and
viper.RegisterAlias does not do what would be expected as ENV handling
is a very special case. This commit fixes up the handling of aliases
specifically for chunk size and minimuim chunk size.

* CLI Option Set
* NEW ENV VAR
* Deprecated ENV Var
* Default Value

For ChunkSize and MinimumChunkSize we now explicitly disallow setting
both --chunk-size and --minimum-chunk-size as CLI options.

If PGET_CHUNK_SIZE is set, we utilize that value except if the cli
option(s) are set.

If PGET_MINIMUM_CHUNK_SIZE is set
* If PGET_CHUNK_SIZE is set to a non-default value a warning is emitted,
  and PGET_CHUNK_SIZE is used..

* If PGET_CHUNK_SIZE is the default value, PGET_MINIMUM_CHUNK_SIZE is
  set as PGET_CHUNK_SIZE and a deprecation warning is emitted.

Additionally, viper calls now all reference config.OptChunkSize