fixed chunk size (#173)
We have a fair amount of complexity (EqualSplit etc) involved in dynamically
splitting files down into chunks based on available concurrency.

We could be less clever and use a fixed chunk size, with a fixed buffer size to
go with it.  This would dovetail nicely with a sync.Pool of fixed-size buffers,
and also with a configuration option to limit the total allocated
RAM (implemented as a semaphore guarding how many buffers can be fetched from
the sync.Pool).

Wire up the --chunk-size option and deprecate --minimum-chunk-size.

* Cleanup Aliasing

Aliases in Cobra and Viper are relatively complex and
viper.RegisterAlias does not do what would be expected as ENV handling
is a very special case. This commit fixes up the handling of aliases
specifically for chunk size and minimuim chunk size.

* CLI Option Set
* NEW ENV VAR
* Deprecated ENV Var
* Default Value

For ChunkSize and MinimumChunkSize we now explicitly disallow setting
both --chunk-size and --minimum-chunk-size as CLI options.

If PGET_CHUNK_SIZE is set, we utilize that value except if the cli
option(s) are set.

If PGET_MINIMUM_CHUNK_SIZE is set
* If PGET_CHUNK_SIZE is set to a non-default value a warning is emitted,
  and PGET_CHUNK_SIZE is used..

* If PGET_CHUNK_SIZE is the default value, PGET_MINIMUM_CHUNK_SIZE is
  set as PGET_CHUNK_SIZE and a deprecation warning is emitted.

Additionally, viper calls now all reference config.OptChunkSize

---------

Co-authored-by: Morgan Fainberg <morgan@replicate.com>