Panic on call to readFrom with emptyBuffer
Ensure it is not possible for downloadBody or other wrappers to
accidently populate data into the emptyBuffer.

This could be solved with an interface, but that interface would likely
need to be quite complex as we rely on the methods on bytes.Buffer.
While we could mirror Cap() Len(), etc, this seems like busy work. It is
easier to broker access to br.buf.ReadFrom through br.readFrom to ensure
we panic in the case of writing to a buffer that has had emptyBuffer
assigned to it.

This ensures we can make the assumption that .Read() will result in EOF
in all scenarios when emptyBuffer has been set on the BufferedReader.