implement Wait() on Strategy to fetch remaining errors (#179)
Currently, we use an errgroup as a semaphore in buffer and consistent hash
modes, but we never inspect the results of sem.Wait() to see if there were any
errors. This results in errors being swallowed.

This implements Wait(), and uses it everywhere we call Fetch().  Now that we
have it, we don't need to store errors on bufferedReader so we can get rid of
that.