Revert "implement Wait() on Strategy to fetch remaining errors (#179)"
This reverts commit c72c38485b2389513b8051912b47e7f63db54641.

We cannot add more to the underlying waitgroup in eg after we call
.Wait(). We're calling .Wait() as part of DownloadFile which could in
some cases result in after .Wait() before return additional additions to
the waitgroup. This is surfaced in multifile mode when the multifile
number of buffers is below a critical threshold allowing for the
downloader to begin a second file before the first file completed.

Closes: #188