Only wait for errGroup once
DownloadFile() was used as a public interface but also internally.  This meant
that it was called multiple times from DownloadFiles(), resulting in calling
Wait on the errgroup multiple times, which caused a panic.

This extracts a common private downloadFile() function which can be called from
both public functions, and ensures we only call Downloader.Wait() once within
each public function.

Fixes #188.