Rearrange concurrency control and error handling
Fixes #188.

This is my second attempt after @tempusfrangit's feedback on #190.

We have too many concurrency primitives, and we have too many ways to handle
errors.

In particular, in both buffer.go and consistent_hashing.go:

- there is an ErrGroup and a workQueue for submitting asynchronous work to, and
  they interact in nontrivial ways
- the `sem` ErrGroup is responsible for imposing concurrency control but it is
  also responsible for collecting errors from asynchronous goroutines.

This commit changes this all:

- the `sem` errgroup is removed
- workQueue is now responsible for concurrency control; it has a limited size
  worker pool.
- nearly all errors that would have been emitted to the ErrGroup are now
  funneled through bufferedReader instead

bufferedReader has has some significant changes:

- ReadFrom() is now Prefetch() and it no longer returns an error
- errors can be injected from outside via recordError() - this is helpful if an
  error that didn't come from the underlying io.Reader is still relevant to the
  consumer (eg if we failed to initiate an http request for a chunk, so we never
  created the response body reader in the first place)

Another big change is that bufferedReader no longer acquires a bufio.Reader from
the bufferPool when created; instead it delays the acquisition until Prefetch()
is called.

This means that it's now safe to create all the bufferedReaders a given Fetch()
call will need ahead of time. Since creating a bufferedReader doesn't allocate a
memory buffer, it won't cause OOM problems.  And this in turn means we can ditch
chanMultiReader and use the standard library io.MultiReader.

workQueue is now a priority work queue.  The priority is because we want to
prioritise finishing existing files over starting new ones; so the first chunk
download happens via submitLow() and all subsequent chunk downloads happen via
submitHigh().

I think long term the right place for concurrency control is the bufferPool - if
the resource we actually care about is RAM, then bufferPool should guard access
to memory buffers with a semaphore.  But we need to be careful of deadlocks if
we do that.