Rewrite everything again
I'm so sorry, I keep doing this.

But I think I have better abstractions now.

The work queue is now responsible for managing buffers, and will only execute
work if there is a buffer available for it.  This means we can limit the size of
the buffer pool but we don't run the risk of deadlock.

bufferedReader has been rewritten to be readerPromise.  It's the promise of an
io.Reader in future.

Instead of using bufio.Reader.Peek() to prefetch data, we prefetch using
io.ReadFull().  This is a really good fit for our use case - we already have an
expected size of the data given by resp.ContentLength; if the data we get is too
short, io.ReadFull returns io.ErrUnexpectedEOF, so we don't need our own
ErrContentLengthMismatch type any more.