Improve Stream Decompression
Instead of implementing the peek reader, wrap everything in bufio.Reader
and lean on it's peek() capabilities. This also allows for a simpler
bytes.HasPrefix use for the magic numbers instead of needing to deal
with the Endian-ness of the magic bytes -- critically this eliminates
the padding for the 48-bit magic bytes header for some compression
types.