Skip global PAXHeaders
Global PAX Headers are a meta-header that applies to subsequent files.
However, in most cases these values are 100% ignorable as the underlying
archive/tar handles merging things. However, the global header stat is
not persisted across headers (as per the spec); notably this is largely
"OK" as most values within the global header are not commonly used or
relevant (i.e. size is unlikely to be relevant as that is file by file
and not global).

We can always add futher PAX Header support if needed, but the reality
is pget is highly optimized for it's use case and doesn't go too far out
of it's way for cases that aren't relevant.