Resume interrupted connections
Resume interrupted connections. This makes the client.Client an
interface as a test seam (so we can easily mock it). Additionally this
now allows pget to resume downloads if a connection is interruped before
the chunk is delivered. This should reduce the potential for
unexpectedEOF errors being bubbled up to the end user.