Address slice issues
Address issues with setting a zero sized slice for the resumable
download. Since we're explicit with download bits and we assume the
buffers can be re-used there is no reason we should not create the
buffers at their actual size.

This commit also cleans up some edge cases with the resume code