Make calls to metrics synchronous (#267)
Metrics are not getting sent in some cases because the metrics don't have enough time to get sent before the program exits. There wasn't enough of a reason for it to be in a WaitGroup since it doesn't really wait for anything, so might as well just send the metrics synchronously.