Always use the range header if it's available (#273)
If we do get a range header, even on a 200, we should use the range header to figure out the total size instead of the content-length. We are making this change because we are now seeing CAIOS returned values of 200s with content lengths that do not match the content range's total size.